+++ src/main/kotlin/org/metalscraps/lostark/item/ItemGrade.kt
@@ -1,6 +1,6 @@
 package org.metalscraps.lostark.item
 
-enum class ItemGrade(val id: Short, val friendlyName: String) {
+enum class ItemGrade(val id: Byte, val friendlyName: String) {
     ALL(-1, "전체"),
     NORMAL(0, "일반"),
     UNCOMMON(1, "고급"),
